Move api module into platform::linux
