--- src/api/mod.rs
@@ -1,5 +1,0 @@
-// TODO: remove this module altogether and move all implementations to `platform`
-
-pub mod dlopen;
-pub mod wayland;
-pub mod x11;
