+++ src/platform/linux/x11/xdisplay.rs
@@ -7,7 +7,7 @@ use std::sync::Mutex;
 use libc;
 
 use super::ffi;
-use api::dlopen;
+use super::super::dlopen;
 
 /// A connection to an X server.
 pub struct XConnection {
